% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnInt}
\alias{dmvnInt}
\title{Evaluates the mutltivariate normal density function}
\usage{
dmvnInt(X, mu, L)
}
\arguments{
\item{X}{arma::mat - rows will be evaluated by the normal density function}

\item{mu}{arma::vec- mean of the distribution}

\item{L}{arma::mat - lower cholesky matrix of the bandwidth matrix H}
}
\description{
Evaluates the mutltivariate normal density function
}
