% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_chol}
\alias{get_chol}
\title{Get the lower cholesky decomposition of  a matrix}
\usage{
get_chol(H)
}
\arguments{
\item{H}{matrix - the matrix to perform cholesky decomposition on}
}
\value{
Lower cholesky matrix of H
}
\description{
Get the lower cholesky decomposition of  a matrix
}
