% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{k_fold_cross_validation_lr}
\alias{k_fold_cross_validation_lr}
\title{Cross validation for the local regression}
\usage{
k_fold_cross_validation_lr(y, x, X, L, k)
}
\arguments{
\item{y}{numeric - outcomes of the data}

\item{x}{matrix - samples from the data}

\item{X}{matrix - feature transforms of the rows of x}

\item{L}{matrix - is the lower cholesky matrix}

\item{k}{numeric - the number of groups to partition the data into for cross-validation.}
}
\value{
the cross-validation error
}
\description{
Cross validation for the local regression
}
