% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_local_pred}
\alias{lm_local_pred}
\title{Train a local least regression model and make a prediction at a single point}
\usage{
lm_local_pred(y, x0, X0, x, X, L)
}
\arguments{
\item{y}{arma::vec - outcomes of the data}

\item{x0}{arma::vec- the point to make a prediction at}

\item{X0}{arma::vec - feature transform of x0}

\item{x}{arma::mat - rows are samples from the data}

\item{X}{arma::mat - feature transform's of rows of x}

\item{L}{arma::mat - Lower Cholesky matrix of the bandwidth matrix}
}
\description{
Train a local least regression model and make a prediction at a single point
}
