% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lm_qr}
\alias{lm_qr}
\title{Computes the least square parameters using QR decomposition}
\usage{
lm_qr(X, y)
}
\arguments{
\item{X}{arma::mat - rows are samples from the data}

\item{y}{arma::mat - outcomes corresponding to the rows of X}
}
\value{
parameters of the least squares estimator
}
\description{
Computes the least square parameters using QR decomposition
}
