\name{lr-package}
\alias{lr-package}
\alias{lr}
\docType{package}
\title{
  A package for local regression
}
\description{
  A package for local regression made with Rcpp. Made for the Statistical Computing 2 unit at University of Bristol COMPASS CDT.
}
\details{
}
\author{
  Conor Newton, conornewton@gmail.com
}
