% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{mse}
\alias{mse}
\title{Computes the mean square error between to vectors}
\usage{
mse(ys1, ys2)
}
\arguments{
\item{ys1}{numeric}

\item{ys2}{numeric}
}
\description{
Computes the mean square error between to vectors
}
